feat: fix issues
